# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Receipt Assistant API",
    "version": "2.0.0-alpha",
    "description": "Double-entry ledger backend. The v1 resources (transactions, postings, accounts, documents) are under active construction — see issues #33 (schema), #35 (transactions API), #36 (accounts API)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Violation": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "code"
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "violations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Violation"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
          },
          "kind": {
            "type": "string",
            "enum": [
              "receipt_image",
              "receipt_email",
              "receipt_pdf",
              "statement_pdf",
              "other"
            ]
          },
          "file_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "mime_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "sha256": {
            "type": "string"
          },
          "ocr_text": {
            "type": [
              "string",
              "null"
            ]
          },
          "extraction_meta": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {}
          },
          "source_ingest_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "kind",
          "file_path",
          "mime_type",
          "sha256",
          "ocr_text",
          "extraction_meta",
          "source_ingest_id",
          "created_at",
          "updated_at"
        ]
      },
...
      "UploadDocumentForm": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          },
          "kind": {
            "type": "string",
            "enum": [
              "receipt_image",
              "receipt_email",
              "receipt_pdf",
              "statement_pdf",
              "other"
            ]
          }
        }
      },
      "CreateDocumentLinkRequest": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
          }
        },
        "required": [
          "transaction_id"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/documents": {
      "post": {
        "summary": "Upload a document (multipart, sha256 content-dedup)",
        "tags": [
          "documents"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadDocumentForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Duplicate upload detected — returns the existing document row",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "201": {
            "description": "Document created",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{id}": {
      "get": {
        "summary": "Get document metadata",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Document",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match matched)"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Hard-delete a document (only if it has no links)",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Document has links — unlink first",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{id}/content": {
      "get": {
        "summary": "Stream document binary content",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Raw file bytes",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{id}/links": {
      "post": {
        "summary": "Link a document to a transaction (idempotent)",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentLinkRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Linked (or already linked — idempotent)"
          },
          "404": {
            "description": "Document or transaction not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{id}/links/{txn_id}": {
      "delete": {
        "summary": "Unlink a document from a transaction",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "txn_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Unlinked"
          },
          "404": {
            "description": "Link not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
...
}